closure_js_library: Add "library_level_checks" attribute
This change adds a new boolean attribute to closure_js_library that
allows the user to opt-out of "library level checks".

These checks --triggered by the named output %{name}_typecheck--, add
a node in the action graph to execute JsChecker against an "i.js" file
that represents an "ABI" of the library javascript.  While this
is normally harmless, in certain circumstances the generated ijs file
is syntactically incorrect and compilation errors against it represent
a false-negative in the overall javascript compile action graph.

This attribute provides a way for users to opt-out of compiling the ijs
file in these circumstances.

See #359