Fix: `ProtoInfo`-based Graph
This changeset fixes gRPC-web downstream in Gust by changing the
build graph to use `ProtoInfo`-based protocol buffer dependencies
for templates, rather than the old-style `rules_closure` aspect.

The build also seems to have sped up in a big way, probably because
of the elimination of the duplicate `closure_proto_library` build
graph (which can now be discarded).

HUGE thanks to @Yannic for help and guidance on this.

Updates:
- [x] Make Java template libraries use `ProtoInfo`
- [x] Make JS template libraries use `ProtoInfo`
- [x] Make Python template libraries use `ProtoInfo`